requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==4.0.2
Babel==2.9.1
billiard==3.6.4.0
Brotli==1.0.9
cached-property==1.5.2
celery==5.1.2
celery-progress==0.1.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.8
django-etc==1.3.1
django-extra-fields==3.0.2
django-import-export==2.6.1
django-model-utils==4.2.0
django-sitemessage==1.3.4
django-summernote==0.8.11.6
django-wkhtmltopdf==3.3.0
djangorestframework==3.13.1
enum34==1.1.10
et-xmlfile==1.1.0
fonttools==4.27.1
gevent==21.12.0
graphene==2.1.9
graphene-django==2.1rc1
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.3
iso8601==0.1.16
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pdfkit==1.0.0
Pillow==8.3.1
poloniex==0.0.17
progress==1.6
promise==2.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.7
pyphen==0.11.0
PySimpleGUI==4.60.1
pytz==2018.7
PyYAML==6.0
qrcode==6.1
redis==4.3.3
reportbro-fpdf==1.7.11
reportbro-lib==2.0.1
reportbro-simpleeval==0.9.11
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
tinycss2==1.1.1
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.1
webencodings==0.5.1
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.1.9
